feat: add float precision service

+++ internal/receive/dto/receive_dto.go
@@ -5,11 +5,11 @@ import "time"
 type SensorData struct {
 	Timestamp time.Time `json:"timestamp"`
 	DeviceID  string    `json:"device_id"`
-	CO        float32   `json:"co"`
-	Humid     float32   `json:"humid"`
-	Temp      float32   `json:"temp"`
-	LPG       float32   `json:"lpg"`
-	Smoke     float32   `json:"smoke"`
+	CO        float64   `json:"co"`
+	Humid     float64   `json:"humid"`
+	Temp      float64   `json:"temp"`
+	LPG       float64   `json:"lpg"`
+	Smoke     float64   `json:"smoke"`
 	Light     bool      `json:"light"`
 	Motion    bool      `json:"motion"`
 }
